display initial framrate and resolution, fix bug that overwrites the settings after being set by presets

+++ ConvertWithMe.UI/Models/SettingsVideo.cs
@@ -17,7 +17,7 @@ namespace ConvertWithMe.UI.Models
         [ObservableProperty]
         private int bitrate;
         [ObservableProperty]
-        private int frameRate;
+        private float frameRate;
         [ObservableProperty]
         private int width;
         [ObservableProperty]
